告警处理

+++ camera/vdi_base/common/adapter/platform/v4l2/src/driver_adapter/src/v4l2_fileformat.cpp
@@ -27,7 +27,7 @@ void HosFileFormat::V4L2GetCurrentFormat(int fd, std::vector<DeviceFormat>& fmtD
     struct v4l2_frmivalenum  fraMival = {};
     constexpr int32_t fmtMax = 50;
     for (int k = 0; k < fmtMax; ++k) {
-        fraMival.index = k;
+        fraMival.index = static_cast<uint32_t>(k);
         fraMival.pixel_format = frmSize.pixel_format;
         fraMival.width = frmSize.discrete.width;
         fraMival.height = frmSize.discrete.height;
